make heartbeat listener async
